+++ src/advanced_ui/checkbox.rs
@@ -193,6 +193,8 @@ impl Element for Checkbox {
                 .value_required(value)?
                 .with_checked(self.checked)
                 .with_enabled(!self.state.disabled())
+                .with_read_only(self.state.read_only())
+                .with_invalid(self.state.invalid())
                 .with_focused(cx.a11y_has_focus(self.id));
         if self.state.can_activate() {
             node = node.with_action(AccessibilityAction::Activate);